requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
PyMySQL==1.1.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-exporter-jaeger==1.21.0
